Allow freely setting project thresholds
Previously the `active` field of the thresholds struct was populated by
checking whether the `threshold` was above `0` or not. However the
`Total Project Threshold` can never be disabled and it is possible for
other thresholds to be at `0` without being disabled too.

To ensure a more accurate representation of the backend state, the
`set_threshold` method instead takes a `Threshold` instance which can be
populated fully by the consumer without any internal restrictions.

See #288.